// src/api/generated/model/questionTemplateCreate.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Amrita ICPC Coding Platform - Backend
 * A backend tool for Amrita ICPC Coding Platform
 * OpenAPI spec version: 1.0.0
 */

export interface QuestionTemplateCreate {
  /** @exclusiveMinimum 0 */
  language_id: number;
  starter_code: string;
  driver_code?: string | null;
  solution_code?: string | null;
}